reliable waiting until end of a message

+++ STM32F0_SIM800_MQTT_Client/inc/SIM808.h
@@ -5,7 +5,7 @@ void simInit(void);
 int simConnect();
 void simSend(const char* data);
 int simReceive(const uint8_t* data, uint8_t length);
-int simAvailable(void);
+uint8_t simAvailable(void);
 int simRead(void);
